+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-Unreleased
-----------
+0.18.0
+------
 - Introduced trailing stop order types
 - Added support for paging of account activity data
 - Added support for specifying reported account activity direction as